Crate problem061

Problem 61 (Cyclical figurate numbers)

Solution detail

We can use a depth-first search, starting with octagonal numbers as the first-level nodes, and having child nodes formed from parent nodes by adding one more polygonal number (meeting the digit constraints) to the set. It is possible to assume that we should start with octagonal numbers because solutions are, by definition, cyclic, so any solution can be rotated to one that starts with an octagonal number.

The search can be sped up by precomputing lists of four-digit polygonal numbers, so that they do not need to be recalculated every time we take a step.

Structs

CyclicPolygonalTree
A structure that can be used for conducting a depth-first search for sets of cyclic polygonal numbers.

CyclicPolygonalTreeStep
A description of a step that can be taken in the search tree.

Functions

answer
Solve the problem, returning the answer as a String

digits_match

main

solve
Find the sum of all cyclic sequences of 4-digit polygonal numbers having the given size.